+++ LEKTOR/LEKTOR_MIKROCONTROLLER/LEKTOR_MIKROCONTROLLER/ToggleSwitchLED.c
@@ -4,7 +4,7 @@
 
  #include "ToggleSwitchLED.h"
 
- void initToggleSwitchLED(char register_, int port)
+ void initToggleSwitchLED(char register_, short int port)
  {
 	if (register_ > 'L' || register_ < 'A' || register_ == 'I')
 	{
